Rename auth components
